% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OrthologueAssignment.R
\name{BuildOrthologues}
\alias{BuildOrthologues}
\title{Main function for Builduing a global useable Orthologue List}
\usage{
BuildOrthologues(GTF.1, GTF.2, species.1, species.2, alignment_type = "global")
}
\arguments{
\item{GTF.1}{path to GTF file for e.g. human genome}

\item{GTF.2}{path to GTF file for e.g. mice genome}

\item{species.1}{name of the first species e.g. human, mouse, zebrafish, more to come...}

\item{species.2}{name of the second species e.g. human, mouse, zebrafish, more to come...}

\item{alignment_type}{Set the type of alignment for orthologue assignment, default is "global" for Needleman-Wunsch, use "local" for Smith-Waterman}
}
\value{
List with uniquely assigned Orthologues between Human and mice
}
\description{
This function uses GTF files for Human and Mice and starts building a Global useable Orthologue List
by assigning unique orthologues to human gene symbols. It uses Ensembl, Uniprot and NCBI databases.
}
\author{
Mariano Ruz Jurado
}
