% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterCells.R
\name{FilterDeadCellsByQuantile}
\alias{FilterDeadCellsByQuantile}
\title{Filter Seurat Object automatically}
\usage{
FilterDeadCellsByQuantile(
  seuratObject,
  lowQuantile = 0.1,
  highQuantile = 0.95,
  maxMito = 0.1
)
}
\arguments{
\item{seuratObject}{a Seurat Object}
}
\value{
filtered seurat object
}
\description{
The number of features and UMIs (nFeature_RNA and nCount_RNA) are automatically calculated for every object by Seurat.
For non-UMI data, nCount_RNA represents the sum of the non-normalized values within a cell
We calculate the percentage of mitochondrial features here and store it in object metadata as \code{percent.mito}.
We use raw count data since this represents non-transformed and non-log-normalized counts
The \% of UMI mapping to MT-features is a common scRNA-seq QC metric.
}
\author{
David John
}
