% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OrthologueAssignment.R
\name{firstup}
\alias{firstup}
\title{function for uppercase the first letter of a character}
\usage{
firstup(x)
}
\arguments{
\item{x}{character}
}
\value{
character with first letter uppercase
}
\description{
function for uppercase the first letter of a character
}
\author{
Mariano Ruz Jurado
}
