% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Importer.R
\name{Importer}
\alias{Importer}
\title{Import Single cell sequencing experiments into Seurat and perform normalisation and scale Data and do a summary of mapping stats}
\usage{
Importer(
  pathway,
  id,
  TenX = TRUE,
  performNormalisation = TRUE,
  performScaling = FALSE,
  performVariableGeneDetection = TRUE,
  FilterCells = TRUE,
  FilterByAbsoluteValues = FALSE,
  ...
)
}
\arguments{
\item{TenX}{Logical for TenX result analysis}

\item{performNormalisation}{Logical if Seurat Normalization should be performed}

\item{performScaling}{Logical if Seurat ScaleData should be performed}

\item{performVariableGeneDetection}{Logical if Seurat FindVariableFeatures should be performed}

\item{FilterCells}{Logical if Cells should be filtered}

\item{FilterByAbsoluteValues}{Logical if Cells should be filtered by manual values, further arguments must be provided for minFeatures,maxFeatures,minCounts,maxCounts,maxMito}

\item{pathways}{A vector of pathways to the cellrancer count output folder (contains barcodes.tsv, genes.tsv, matrix.mtx)}

\item{ids}{Vector of strings that are assigned to the concordant cells}
}
\value{
Merged seurat object
}
\description{
Import Single cell sequencing experiments into Seurat and perform normalisation and scale Data and do a summary of mapping stats
}
\author{
David John & Mariano Ruz Jurado
}
