% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OrthologueAssignment.R
\name{protein.matching}
\alias{protein.matching}
\title{protein sequence matching alignment}
\usage{
protein.matching(mGene, replacement, OrthologueList_allHuman)
}
\arguments{
\item{mGene}{Gene which will be matched against its possible orthologues in replacement}

\item{replacement}{vector containing possible orthologues found by biomaRt}

\item{OrthologueList_allHuman}{global orthologuelist which contains all Human genes and is filled with mouse orthologues}
}
\description{
Reads in protein sequences from Uniprot for possible orthologues and calculates a score by pairwise alignment comparison
}
\author{
Mariano Ruz Jurado
}
