% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SubsetObjects.R
\name{RenameGenesSeurat}
\alias{RenameGenesSeurat}
\title{Renaming of Mice Objects with the found human names from SubsetObjects}
\usage{
RenameGenesSeurat(ObjList, newnames)
}
\arguments{
\item{ObjList}{List with Seurat object to subset with orthologues names}

\item{newnames}{List with vector with new names generated by SubsetObjects function}
}
\value{
Seurat object containing subsetted object with human orthologue names
}
\description{
Renaming of Mice Objects with the found human names from SubsetObjects
}
\author{
Mariano Ruz Jurado
}
