% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SubsetObjects.R
\name{RenameGenesSeuratv5}
\alias{RenameGenesSeuratv5}
\title{Renaming of Objects with the found names from BuildOrthologues, works for Seuv5 objects}
\usage{
RenameGenesSeuratv5(ObjList, newnames)
}
\arguments{
\item{ObjList}{List with Seuratv5 objects to subset with orthologues names}

\item{newnames}{List with vector with new names generated by SubsetObjects function}
}
\value{
Seurat object containing subsetted object with human orthologue names, ready for Integration
}
\description{
Renaming of Objects with the found names from BuildOrthologues, works for Seuv5 objects
}
\author{
Mariano Ruz Jurado
}
