% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OrthologueAssignment.R
\name{returnOrthologueslist}
\alias{returnOrthologueslist}
\title{ensembl database orthologue setting using biomart function, converting mouse genes into human genes}
\usage{
returnOrthologueslist(GeneNames)
}
\arguments{
\item{GeneNames}{Vector of Human gene names}
}
\value{
List with converted Names based on ensembl database entries
}
\description{
ensembl database orthologue setting using biomart function, converting mouse genes into human genes
}
\author{
Mariano Ruz Jurado
}
