% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SubsetObjects.R
\name{SubsetObjects}
\alias{SubsetObjects}
\title{Subsets objects by orthologues found in the global list and gives Human gene names for Mouse objects}
\usage{
SubsetObjects(OrthologueList, SeuratObjectList.mice, SeuratObjectList.human)
}
\arguments{
\item{OrthologueList}{previous mad elist with 1 to 1 ortholgue assignment}

\item{SeuratObjectList.mice}{list with mouse objects to subset by found orthologues}

\item{SeuratObjectList.human}{list with human seurat objects to subset by orthologues}
}
\value{
list containing sublists with subsetted objects
}
\description{
Subsets objects by orthologues found in the global list and gives Human gene names for Mouse objects
}
\author{
Mariano Ruz Jurado
}
