% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummarizeMapping.R
\name{SummarizeMapping}
\alias{SummarizeMapping}
\title{Summarize mapping statistics
This function reads in the summary file about the mapping with CellRanger or Starsolo and gives warnings if values surpass a threshold}
\usage{
SummarizeMapping(pathways, ids)
}
\arguments{
\item{pathway}{A vector of pathways to the cellranger/star solo count result}

\item{id}{Vector of strings that are assigned to the concordant cells}
}
\value{
a list with summary about mapping results
}
\description{
Summarize mapping statistics
This function reads in the summary file about the mapping with CellRanger or Starsolo and gives warnings if values surpass a threshold
}
\author{
Mariano Ruz Jurado
}
