% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OrthologueAssignment.R
\name{BuildOrthologues}
\alias{BuildOrthologues}
\title{Main function for Builduing a global useable Orthologue List}
\usage{
BuildOrthologues(GTF.human, GTF.mice)
}
\arguments{
\item{GTF.human}{path to GTF file for human genome}

\item{GTF.mice}{path to GTF file for mice genome}
}
\value{
List with uniquely assigned Orthologues between Human and mice
}
\description{
This function uses GTF files for Human and Mice and starts building a Global useable Orthologue List
by assigning unique orthologues to human gene symbols. It uses Ensembl, Uniprot and NCBI databases.
}
\author{
Mariano Ruz Jurado
}
