% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Importer.R
\name{Importer}
\alias{Importer}
\title{Import Single cell sequencing experiments into Seurat and perform normalisation and scale Data}
\usage{
Importer(
  pathways,
  ids,
  TenX = TRUE,
  performNormalisation = TRUE,
  performScaling = FALSE,
  performVariableGeneDetection = TRUE,
  FilterCells = TRUE,
  FilterByAbsoluteValues = FALSE,
  ...
)
}
\arguments{
\item{pathways}{A vector of pathways to count output folder of mapping tool (for 10X: Directory  contains barcodes.tsv, genes.tsv, matrix.mtx)}

\item{ids}{Vector of strings that are assigned to the concordant cells}

\item{TenX}{Logical for TenX result analysis}

\item{performNormalisation}{Logical if Seurat Normalization should be performed}

\item{performScaling}{Logical if Seurat ScaleData should be performed}

\item{performVariableGeneDetection}{Logical if Seurat FindVariableFeatures should be performed}

\item{FilterCells}{Logical if Cells should be filtered}

\item{FilterByAbsoluteValues}{Logical if Cells should be filtered by manual values, further arguments must be provided for minFeatures,maxFeatures,minCounts,maxCounts,maxMito}
}
\value{
Merged seurat object
}
\description{
Import Single cell sequencing experiments into Seurat and perform normalisation and scale Data
}
\author{
David John & Mariano Ruz Jurado
}
