% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SubsetObjects.R
\name{IntegrateObjects}
\alias{IntegrateObjects}
\title{Wrapper function for SubsetObject and RenameGenesSeurat as well as the Seurat integration steps}
\usage{
IntegrateObjects(
  SeuratObjectList.species.1,
  SeuratObjectList.species.2,
  OrthologueList
)
}
\arguments{
\item{SeuratObjectList.species.1}{list with objects to subset by found orthologues}

\item{SeuratObjectList.species.2}{list with seurat objects to subset by orthologues}

\item{OrthologueList}{previously made list with 1 to 1 orthologue assignment}
}
\value{
Integrated Human/Mouse Seurat object with Human Nomenclature
}
\description{
Wrapper function for SubsetObject and RenameGenesSeurat as well as the Seurat integration steps
}
\author{
Mariano Ruz Jurado
}
