% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SubsetObjects.R
\name{IntegrateObjects}
\alias{IntegrateObjects}
\title{Wrapper function for SubsetObject and RenameGenesSeurat as well as the Seurat integration steps}
\usage{
IntegrateObjects(SeuratObjectList.human, SeuratObjectList.mice, OrthologueList)
}
\arguments{
\item{SeuratObjectList.human}{list with human seurat objects to subset by orthologues}

\item{SeuratObjectList.mice}{list with mouse objects to subset by found orthologues}

\item{OrthologueList}{previously made list with 1 to 1 ortholgue assignment}
}
\value{
Integrated Human/Mouse Seurat object with Human Nomenclature
}
\description{
Wrapper function for SubsetObject and RenameGenesSeurat as well as the Seurat integration steps
}
\author{
Mariano Ruz Jurado
}
