% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SubsetObjects.R
\name{SubsetObjects}
\alias{SubsetObjects}
\title{Subsets objects by orthologues found in the global list and gives Human gene names for Mouse objects}
\usage{
SubsetObjects(
  OrthologueList,
  SeuratObjectList.species.1,
  SeuratObjectList.species.2,
  species.1,
  species.2
)
}
\arguments{
\item{OrthologueList}{previous mad elist with 1 to 1 orthologue assignment}

\item{SeuratObjectList.species.1}{list with objects of species 1 to subset by found orthologues}

\item{SeuratObjectList.species.2}{list with seurat objects of species 2 to subset by orthologues}

\item{species.1}{name of the species, ("human","mouse", "zebrafish")}

\item{species.2}{name of the species, ("human","mouse", "zebrafish")}
}
\value{
list containing sublists with subsetted objects
}
\description{
Subsets objects by orthologues found in the global list and gives Human gene names for Mouse objects
}
\author{
Mariano Ruz Jurado
}
